Replace Apache Commons Math3 with Maven dependency
This will obviate the need to maintain code that is already maintained
by the Apache Software Project. Instead, we will easily be able to
import the latest version of the library via the dependency manager
(Maven). Less code in the codebase means that people can concentrate and
more easily learn that which is unique to this project.